Move customVirtualEnvLocator.ts
